feat: Leave backticks unencoded in query strings where possible
Addresses #1397.

When the user provides (part of) their query as a string, we leave any
unencoded backticks in the string exactly as is, trusting that the user
wanted them that way.

When the user calls req.query(Object), however, and one of the
keys/values in Object has an unencoded backtick, we continue to encode
that.

We could choose not to encode that as well, and that would make us a
bit more consistent from one POV. But, it would make things less
consistent between the Node and browser versions of superagent, because
the browser version serializes query objects with the
`application/x-www-form-urlencoded` serializer, which is unambiguous
that backticks must be encoded.

Moreover, leaving backticks unencoded doesn’t buy us anything
functionally, because a fundamental limit of the `req.query(Object)`
format is that the user can't indicate "a backtick meant as a
delimiter" vs. "a backtick meant as data”. This limitation becomes
apparent when you consider what happens if the user tries to provide an
already encoded backtick in their query object, in an attempt to
distinguish these two cases, namely: the % sign gets encoded again to
%25, like any other object character needing encoding. And there's no
way around that/that’s how it should be. (Consider that all other
potential delimiter characters that can appear unencoded have the same
behavior. There’s just no way, in the object form, to know what the
user wants.)

Given that leaving backticks unencoded in the object form wouldn’t
enable new functionality (as it does in the string form), I’ve decided
here to encode them, for consistency with the browser superagent.

The new test in `test/client/serialize.js` is to recognize that the
serialize function tested in that file is a serializer for
`application/x-www-form-urlencoded`, and its spec is unambiguous that
backticks there get percent-encoded, regardless of what we do elsewhere.